// backend/types.ts
export interface inputType {
    businessField : String, // תחום עיסוק
    businessFieldDetails: String, // תיאור תחום העיסוק
    businessName: String, // שם העסק
    businessType: String, // סוג העסק
    serviceAreas: String, // אזורי שירות
    serviceDescription: String, // תיאור השירות
    socialMediaAccounts: String, // חשבונות רשתות חברתיות
    contactInfo:String, // פרטי יצירת קשר
    location: String, // מיקום
    callToAction: String, // קריאה לפעולה
    ageGroup: String, // קבוצת גיל
    targetAudience: String, // קהל יעד
    yearOfExperience: String, // שנות ניסיון
}